requirements.txt
yt-dlp>=2023.11.16
gspread>=5.12.4
oauth2client>=4.1.3
python-dotenv>=1.0.0
pycryptodomex>=3.19.0 